split repository code
splits the code in repository/repository for cleanliness.
No logic changes, just code movement.  This is prep
for the nest PR.
